requirements.txt
flask
python-dotenv

langchain
langchain-core
langchain-google-genai

google-generativeai
google-auth
cryptography
